backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
black==26.1.0
boto3==1.42.39
botocore==1.42.39
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
# emergentintegrations==0.1.0  # custom local package
et_xmlfile==2.0.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.3
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
# google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth>=2.29.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
# google-genai==1.61.0
# google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
librt==0.7.8
litellm==1.80.0
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mpmath==1.3.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
notion-client==2.7.0
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.24.1
openai==1.99.9
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pathspec==1.0.4
pdf2image==1.17.0
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
pluggy==1.6.0
pooch==1.9.0
propcache==0.4.1
proto-plus==1.27.0
protobuf>=5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
PyMatting==1.1.15
pymongo==4.5.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
regex==2026.1.15
rembg==2.0.72
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
scikit-image==0.26.0
scipy==1.17.0
shellingham==1.5.4
signnow_python_sdk==2.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.37.2
stripe==14.3.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2026.1.28
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.2
twilio==9.10.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
# tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
